Migrate to Hibernate

+++ src/main/java/kz/zhanbolat/spring/repository/EventRepository.java
@@ -7,6 +7,6 @@ import java.util.Optional;
 
 public interface EventRepository {
     boolean createEvent(Event event);
-    Optional<Event> getEvent(int id);
+    Optional<Event> getEvent(Long id);
     List<Event> getEvents();
 }
